#77 stat mode didn't push during last commit...

+++ src/main/java/Core/Game.java
@@ -2,9 +2,8 @@ package Core;
 
 import Card.*;
 import Exceptions.WondersException;
-import Player.Player;
-import Utility.RecapScore;
 import Player.*;
+import Utility.RecapScore;
 import Utility.Utilities;
 import Utility.Writer;
 
@@ -37,13 +36,57 @@ public class Game {
 
     public static void main(String[] args) throws ParseException, IOException, WondersException {
 
-        Writer.init(true);
-        Game game = new Game(3);
-        while(game.state != GameState.EXIT)
-            game.process();
-        Writer.close();
-        Utilities.displayGameOutput();
-        Writer.deleteFile();
+        int nbJoueur = 3;
+        String typePartie  = MODE_PARTIE;
+
+        if(typePartie.equals(MODE_PARTIE))
+        {
+            Game game = new Game(3);
+            Writer.init(true);
+            while(game.state != GameState.EXIT)
+                game.process();
+            Writer.close();
+            Utilities.displayGameOutput();
+            Writer.deleteFile();
+        }
+        else if (typePartie.equals(MODE_STATS))
+        {
+            Writer.init(false);
+            RecapScore[] recapScores = new RecapScore[3];
+            for(int i=0; i<recapScores.length ; i++)
+            {
+                recapScores[i] = new RecapScore();
+            }
+            for(int i=0 ; i<NOMBRE_PARTIES_STATS ; i++)
+            {
+                Game game = new Game(3);
+                while(game.state != GameState.EXIT)
+                    game.process();
+                RecapScore[] scoresTmp = game.displayPlayersRanking();
+                for(int j=0; j<recapScores.length ; j++)
+                {
+                    recapScores[j].addRecap(scoresTmp[j]);
+                }
+            }
+
+
+            for(int i=0; i<playersArray.size() ; i++)
+            {
+                recapScores[i].processAvgScore(NOMBRE_PARTIES_STATS);
+                double victoires = (recapScores[i].getNbVictory()/(double)NOMBRE_PARTIES_STATS)*100;
+                String joueur= playersArray.get(i).toString();
+                System.out.println(joueur +" obtient un score moyen de "+recapScores[i].getAvgScore());
+                System.out.println(joueur +" gagne dans "+ victoires +"% des parties");
+
+            }
+        }
+        else
+        {
+            throw new RuntimeException("Mode inexistant.");
+        }
+
+
+
 
     }
 
@@ -68,7 +111,7 @@ public class Game {
 
     public void initPlayers() {
         for (int i = 0; i < players; i++)
-            this.playersArray.add(new Player("Bot" + i));
+            this.playersArray.add(new DumbPlayer("Bot" + i));
         for (int i = 0; i < players; i++) {
             Player prevPlayer, nextPlayer;
             if (i > 0) {
@@ -232,8 +275,7 @@ public class Game {
 
         ArrayList<Player> players = this.getPlayersArray();
         Player tmpWinner = players.get(0);
-        RecapScore[] playersScores = new RecapScore[getPlayers()];
-        int offset = 0;
+        RecapScore[] playerScores = new RecapScore[players.size()];
         for(Player p : players) {
             Writer.write(p.getName() + " :  " + p.getCoins() + "coins");
             Writer.write(p.getName() + " :  " + p.getSciencePoint() + "science points");
@@ -242,11 +284,14 @@ public class Game {
                 (p.computeScore() == tmpWinner.computeScore() && p.getCoins() > tmpWinner.getCoins())) {
                 tmpWinner = p;
             }
-
         }
 
+        for(int i = 0; i < players.size();i++)
+            playerScores[i] = new RecapScore(players.get(i),players.get(i).equals(tmpWinner));
 
         Writer.write(tmpWinner.getName() + " won the game with " + tmpWinner.computeScore() + " points !");
+
+        return playerScores;
 
     }
